+++ SharpSource/SharpSource/Utilities/DiagnosticId.cs
@@ -65,4 +65,5 @@ public static class DiagnosticId
     public const string ValueTaskAwaitedMultipleTimes = "SS063";
     public const string UnnecessaryToStringOnSpan = "SS064";
     public const string LoggerMessageAttribute = "SS065";
+    public const string DisposableFieldIsNotDisposed = "SS066";
 }